refactor: extract buildDefinitionText to eliminate duplicated truncation logic
buildDefinition (ud-bot) and buildMessageText (inline-results) were identical
after both switched to templates.definition. Extract the shared logic to
buildDefinitionText in templates.ts — the natural owner since it wraps the
template rendering. Both callers are reduced to a one-liner.
Rename inline-results.test.ts → templates.test.ts to match what it tests.
