+++ src/features/definitions/templates.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { buildMessageText } from "./inline-results";
+import { buildDefinitionText } from "./templates";
 import type { UdDefinition } from "./definition";
 
 // ─── helpers ────────────────────────────────────────────────────────────────
@@ -36,14 +36,14 @@ const TELEGRAM_LIMIT = 4096;
 
 describe("buildMessageText", () => {
   it("returns the full message when both fields fit", () => {
-    const result = buildMessageText(makeDefinition());
+    const result = buildDefinitionText(makeDefinition());
     expect(result.length).toBeLessThanOrEqual(TELEGRAM_LIMIT);
     expect(result).not.toContain("Read more");
   });
 
   it("does not truncate when a long definition still fits within the total limit", () => {
     // ~1500 chars def + short example — well under 4096
-    const result = buildMessageText(makeDefinition({ formattedDefinition: repeat("word ", 300) }));
+    const result = buildDefinitionText(makeDefinition({ formattedDefinition: repeat("word ", 300) }));
     expect(result).not.toContain("Read more");
     expect(result.length).toBeLessThanOrEqual(TELEGRAM_LIMIT);
   });
@@ -51,7 +51,7 @@ describe("buildMessageText", () => {
   it("truncates only the definition when it alone pushes the total over the limit", () => {
     const longDefinition = repeat("word ", 900); // ~4500 chars
     const shortExample = "<i>short</i>";
-    const result = buildMessageText(makeDefinition({ formattedDefinition: longDefinition, formattedExample: shortExample }));
+    const result = buildDefinitionText(makeDefinition({ formattedDefinition: longDefinition, formattedExample: shortExample }));
     expect(result.length).toBeLessThanOrEqual(TELEGRAM_LIMIT);
     expect(result).toContain("Read more");
     // The short example must appear verbatim — it was not truncated
@@ -61,7 +61,7 @@ describe("buildMessageText", () => {
   it("truncates only the example when it alone pushes the total over the limit", () => {
     const shortDefinition = "a short definition";
     const longExample = `<i>${repeat("word ", 900)}</i>`;
-    const result = buildMessageText(makeDefinition({ formattedDefinition: shortDefinition, formattedExample: longExample }));
+    const result = buildDefinitionText(makeDefinition({ formattedDefinition: shortDefinition, formattedExample: longExample }));
     expect(result.length).toBeLessThanOrEqual(TELEGRAM_LIMIT);
     expect(result).toContain("Read more");
     // The short definition must appear verbatim — it was not truncated
@@ -71,7 +71,7 @@ describe("buildMessageText", () => {
   it("truncates both fields when both are very long", () => {
     const longDefinition = repeat("word ", 1000);
     const longExample = `<i>${repeat("word ", 1000)}</i>`;
-    const result = buildMessageText(makeDefinition({ formattedDefinition: longDefinition, formattedExample: longExample }));
+    const result = buildDefinitionText(makeDefinition({ formattedDefinition: longDefinition, formattedExample: longExample }));
     expect(result.length).toBeLessThanOrEqual(TELEGRAM_LIMIT);
     expect(result.split("Read more").length - 1).toBe(2);
   });
@@ -79,7 +79,7 @@ describe("buildMessageText", () => {
   it("never exceeds TELEGRAM_LIMIT even with maximum-length inputs", () => {
     const maxDefinition = repeat("a ", 3000);
     const maxExample = `<i>${repeat("b ", 3000)}</i>`;
-    const result = buildMessageText(
+    const result = buildDefinitionText(
       makeDefinition({ word: repeat("w", 50), formattedDefinition: maxDefinition, formattedExample: maxExample }),
     );
     expect(result.length).toBeLessThanOrEqual(TELEGRAM_LIMIT);
@@ -88,7 +88,7 @@ describe("buildMessageText", () => {
   it("does not produce a broken HTML tag when truncation falls mid-tag", () => {
     const prefix = repeat("word ", 390);
     const definitionWithTag = `${prefix}<a href="https://urbandictionary.com/define.php?term=something-very-long">linked word</a>`;
-    const result = buildMessageText(makeDefinition({ formattedDefinition: definitionWithTag }));
+    const result = buildDefinitionText(makeDefinition({ formattedDefinition: definitionWithTag }));
     expect(result).not.toMatch(/<[^>]*$/m);
     expect(result.length).toBeLessThanOrEqual(TELEGRAM_LIMIT);
   });
@@ -96,7 +96,7 @@ describe("buildMessageText", () => {
   it("includes the permalink in the Read more link", () => {
     const permalink = "https://www.urbandictionary.com/define.php?term=bong";
     const longDefinition = repeat("word ", 1000);
-    const result = buildMessageText(makeDefinition({ formattedDefinition: longDefinition, permalink }));
+    const result = buildDefinitionText(makeDefinition({ formattedDefinition: longDefinition, permalink }));
     expect(result).toContain(`href="${permalink}"`);
   });
 });
